# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Masafah API Documentation",
        "description": "Api description...",
        "version": "1.0.0"
    },
    "paths": {
...
        "/~tvavisa/masafah/public/api/company/addToWallet": {
            "post": {
                "tags": [
                    "Company Wallet"
                ],
                "summary": "Add to company's wallet",
                "operationId": "addToWallet",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Update profile body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "wallet_id": {
                                    "description": "Company Wallet ID - *(Required)",
                                    "type": "integer",
                                    "example": 34
                                },
                                "amount": {
                                    "description": "Amount to be added - *(Required)",
                                    "type": "double",
                                    "example": 50
                                },
                                "isOffer": {
                                    "description": "Amount added to wallet is under offers - *(Required)",
                                    "type": "boolean",
                                    "example": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
        "/~tvavisa/masafah/public/api/company/getWalletOffers": {
            "get": {
                "tags": [
                    "Company Wallet"
                ],
                "summary": "Get Wallet offers",
                "operationId": "getWalletOffers",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    }
                }
            }
        },
        "/~tvavisa/masafah/public/api/company/getWalletDetails": {
            "get": {
                "tags": [
                    "Company Wallet"
                ],
                "summary": "Get Wallet Details",
                "operationId": "getWalletDetails",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    }
                }
            }
        },
...
    },
...
}